requirements.txt
beautifulsoup4
boto3
mysql_connector_repackaged
numpy
pandas
PyMySQL
SQLAlchemy
streamlit
mysql_connector_repackaged
mysql-connector-python
